Refactor pipeline structure with helpers and steps packages
