+++ server/steps/transcribe.py
@@ -1,36 +1,34 @@
 from faster_whisper import WhisperModel
-import time
+
+from helpers.timing import Timer
+
 
 def transcribe_audio(file_path, model_size="medium"):
-    """
-    Transcribe an audio file using faster_whisper.
-    Returns a dict with transcription text, segments, and timing info.
-    """
-    start_time = time.time()
-    model = WhisperModel(model_size, device="auto")  # or "cuda" / "metal" if you use those backends
-    segments, info = model.transcribe(file_path)
-    stop_time = time.time()
+    """Transcribe an audio file using faster_whisper."""
+    with Timer() as t:
+        model = WhisperModel(model_size, device="auto")
+        segments, info = model.transcribe(file_path)
     text = "".join([s.text for s in segments])
     segment_list = [
         {
             "start": s.start,
             "end": s.end,
-            "text": s.text
+            "text": s.text,
         }
         for s in segments
     ]
     timing = {
-        "start_time": start_time,
-        "stop_time": stop_time,
-        "total_time": stop_time - start_time
+        "start_time": t.start_time,
+        "stop_time": t.stop_time,
+        "total_time": t.elapsed,
     }
     return {
         "text": text,
         "segments": segment_list,
-        "timing": timing
+        "timing": timing,
     }
 
-# Example usage
+
 if __name__ == "__main__":
     audio_path = "kfaf1.mp3"
     model_name = "large-v3-turbo"
@@ -40,4 +38,4 @@ if __name__ == "__main__":
     print("\nFirst 12 segments:")
     print(result["segments"][:12])
     print("\nTiming info:")
-    print(result["timing"])
+    print(result["timing"])
